requirements.txt
absl-py==1.4.0
astunparse==1.6.3
attrs==22.2.0
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.0.1
chex==0.1.6
dm-tree==0.1.8
exceptiongroup==1.1.0
execnet==1.9.0
flatbuffers==23.1.21
gast==0.4.0
google-auth==2.16.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
graphviz==0.20.1
grpcio==1.51.1
h5py==3.8.0
idna==3.4
iniconfig==2.0.0
jax==0.4.4
jaxlib==0.4.4
keras==2.11.0
libclang==15.0.6.1
liboqs-python==0.7.2
Markdown==3.4.1
MarkupSafe==2.1.2
networkx==3.0
numpy==1.24.2
oauthlib==3.2.2
opt-einsum==3.3.0
optax==0.1.4
packaging==23.0
pluggy==1.0.0
protobuf==3.19.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pytest==7.2.1
pytest-xdist==3.2.0
PyYAML==6.0
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
scipy==1.10.0
six==1.16.0
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorcircuit==0.7.0
tensorflow==2.11.0
tensorflow-estimator==2.11.0
tensorflow-io-gcs-filesystem==0.30.0
tensornetwork==0.4.6
termcolor==2.2.0
tomli==2.0.1
toolz==0.12.0
tqdm==4.64.1
typing_extensions==4.5.0
urllib3==1.26.14
Werkzeug==2.2.3
wrapt==1.14.1
